refactored the files
