+++ frontend/App.js
@@ -6,7 +6,7 @@ import Homepage from "./screens/Homepage";
 import BrandorInfluencer from "./screens/BrandorInfluencer";
 import BrandRegistrationForm from "./screens/BrandRegistrationForm";
 import OtpVerification from "./screens/OtpVerification";
-import LoginPage from "./screens/LoginPage";
+import LoginPage from "./screens/login/LoginPage";
 import FriendRequestPage from "./screens/FriendRequestPage";
 import AccountCreatedSuccessfullyNoti from "./screens/AccountCreatedSuccessfullyNoti";
 import FailedPaymentStatus from "./screens/FailedPaymentStatus";
@@ -17,7 +17,7 @@ import ChatInterface from "./screens/ChatInterface";
 import UserProfile from "./screens/UserProfile";
 import InfluencerRegistrationForm from "./screens/InfluencerRegistrationForm";
 import PlanChooseInterface from "./screens/PlanChooseInterface";
-import LoginPageBrand from "./screens/LoginPageBrands";
+import LoginPageBrand from "./screens/login/LoginPageBrands";
 import BrandAssosciated from "./screens/BrandsAssosciated";
 import BrandProfile from "./screens/BrandProfile";
 import AdminPanel from "./screens/AdminPanel";
